refactor(@calculate@add): move classes to the strategy package
